requirements.txt
customtkinter
langchain
langchain-ollama
langchain-community
langchain-text-splitters
chromadb
pypdf
python-dotenv